deploy/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
idna==2.7
ndg-httpsclient==0.5.1
pyasn1==0.4.4
pycparser==2.18
pyOpenSSL==18.0.0
requests==2.19.1
scaleway-sdk==1.6.0
six==1.11.0
slumber==0.7.1
urllib3==1.23
